+++ cmd/tvault/cmd/lock.go
@@ -1,8 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-
 	"github.com/abdul-hamid-achik/tinyvault/internal/vault"
 
 	"github.com/spf13/cobra"
@@ -23,7 +21,7 @@ func runLock(_ *cobra.Command, _ []string) error {
 	dir := getVaultDir()
 	v, err := vault.Open(dir)
 	if err != nil {
-		return fmt.Errorf("vault not found at %s, run 'tvault init' first", dir)
+		return wrapVaultOpenErr(dir, err)
 	}
 	defer v.Close()