+++ cmd/tvault/cmd/rotate.go
@@ -34,7 +34,7 @@ func runKeyRotate(_ *cobra.Command, _ []string) error {
 	dir := getVaultDir()
 	v, err := vault.Open(dir)
 	if err != nil {
-		return fmt.Errorf("vault not found at %s, run 'tvault init' first", dir)
+		return wrapVaultOpenErr(dir, err)
 	}
 	defer v.Close()