+++ cmd/tvault/cmd/unlock.go
@@ -26,7 +26,7 @@ func runUnlock(_ *cobra.Command, _ []string) error {
 	dir := getVaultDir()
 	v, err := vault.Open(dir)
 	if err != nil {
-		return fmt.Errorf("vault not found at %s, run 'tvault init' first", dir)
+		return wrapVaultOpenErr(dir, err)
 	}
 	defer v.Close()